score updates

+++ GamePanel.java
@@ -14,15 +14,16 @@ public class GamePanel extends JPanel {
   private BufferedImage bomb, tree, gift;
   private Game game;
   private ChristmasTree ct;
-  private int mouse;
-  private final int treeY = 450;
+  private int mouse; private int score;
+  private final int treeY = 440; // 450 + 25
 
   public GamePanel(Game game) {
     this.game = game;
     this.ct = game.getTree();
     this.ct.setLocation(0, treeY);
-    timer = new Timer(500, new ActionListener() { // timer with 150 millisecond delay
+    timer = new Timer(20, new ActionListener() { // timer with 150 millisecond delay
       public void actionPerformed(ActionEvent e) {
+        if (game.win() || game.lose()) game.endGame();
         repaint();
       }
     });
@@ -43,23 +44,34 @@ public class GamePanel extends JPanel {
   }
 
   public Dimension getPreferredSize() {
-    return new Dimension(500, 750);
+    return new Dimension(500, 750); // 250 - 160, 375 - 160
   }
 
   protected void paintComponent(Graphics g) {
-    super.paintComponent(g);
-    game.drop();
-    Vector<Item> active = game.getActive();
+    if (game.didEnd()) {
+      try {
+        BufferedImage image = ImageIO.read(new File("gameOver.png"));
+        if (game.win()) image = ImageIO.read(new File("win.png"));
 
-    ct.setLocation(mouse, treeY);
-    g.drawImage(tree, mouse, treeY, this);
-    //treeLabel.setLocation(mouse, treeY);
+        g.drawImage(image, 90, 215, this);
+      } catch (IOException ex) {
+        System.out.println(ex);
+      }
+    } else {
+      super.paintComponent(g);
+      game.drop();
+      LinkedList<Item> active = game.getActive();
 
-    for (Item item : active) {
-      if (item.isGift()) {
-        g.drawImage(gift, item.getX(), item.getY(), this);
-      } else {
-        g.drawImage(bomb, item.getX(), item.getY(), this);
+      ct.setLocation(mouse, treeY);
+      g.drawImage(tree, mouse, treeY, this);
+
+      for (Item item : active) {
+        // System.out.println("item at: " + item.getX() + ", " + item.getY());
+        if (item.isGift()) {
+          g.drawImage(gift, item.getX(), item.getY(), this);
+        } else {
+          g.drawImage(bomb, item.getX(), item.getY(), this);
+        }
       }
     }
   }
@@ -68,9 +80,7 @@ public class GamePanel extends JPanel {
     public void mouseMoved(MouseEvent event) {
       int x = event.getX() - 150 > 500 ? 500 : event.getX() - 155;
       mouse = x;
-      //mouse = event.getX();
-      System.out.println("Mouse: [" +event.getX() +", "+event.getY()+"] Tree: "+mouse);
-      //System.out.println("mouse:" + event.getX() + "  " + "tree: " + mouse);
+      //System.out.println("tree at: " + ct.getX() + ", " + ct.getY());
     }
   }
 }
